Toolbar multiple item issue on 64 bit - issue 11644
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40259 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Additions/GNUstepGUI/GSToolbarView.h
@@ -60,8 +60,8 @@ typedef enum {
   float _heightFromLayout;
 }
 
-+ (int) draggedItemIndex;
-+ (void) setDraggedItemIndex:(int)sourceIndex;
++ (NSInteger) draggedItemIndex;
++ (void) setDraggedItemIndex:(NSInteger)sourceIndex;
 
 - (id) initWithFrame: (NSRect)frame;
